Use Jackson2JsonMessageConverter and refactor a little bit
