+++ transport_service_commands/example_events.md
@@ -1,0 +1,13 @@
+# Flight
+
+**Create flight**
+```json
+{
+  "departureDate": "2023-04-24T14:15",
+  "departureCountry": "Spain",
+  "departureCity": "Barcelona",
+  "arrivalDate": "2023-04-24T18:15",
+  "arrivalCountry": "Greece",
+  "arrivalCity": "Athens"
+}
+```
