+++ transport_service_commands/pom.xml
@@ -59,7 +59,7 @@
 		<dependency>
 			<groupId>com.fasterxml.jackson.core</groupId>
 			<artifactId>jackson-databind</artifactId>
-			<version>2.13.0</version>
+			<version>2.15.0</version>
 		</dependency>
 
 		<dependency>
