+++ transport_service_commands/src/main/java/pl/edu/pg/transport/entity/Transport.java
@@ -1,13 +1,6 @@
-package pl.edu.pg.transport;
+package pl.edu.pg.transport.entity;
 
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
+import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
