+++ transport_service_queries/example_events.md
@@ -1,0 +1,16 @@
+# Flight
+
+**Get flights**
+```json
+{
+  "source": "testSource"
+}
+```
+
+**Get flight details**
+```json
+{
+  "hotelId": 1,
+  "source": "testSource"
+}
+```
