--- transport_service_queries/src/main/java/pl/edu/pg/transport/GetFlightDetailsEvent.java
@@ -1,14 +1,0 @@
-package pl.edu.pg.transport;
-
-public class GetFlightDetailsEvent {
-    private long id;
-    private String source;
-
-    public long getId() {
-        return id;
-    }
-
-    public String getSource() {
-        return source;
-    }
-}
