+++ transport_service_queries/src/main/java/pl/edu/pg/transport/event/GetFlightDetailsEvent.java
@@ -1,0 +1,9 @@
+package pl.edu.pg.transport.event;
+
+public class GetFlightDetailsEvent extends GenericEvent {
+    private long id;
+
+    public long getId() {
+        return id;
+    }
+}
