+++ transport_service_queries/src/main/java/pl/edu/pg/transport/event/GetFlightsEvent.java
@@ -1,0 +1,4 @@
+package pl.edu.pg.transport.event;
+
+public class GetFlightsEvent extends GenericEvent {
+}
